refactor the header
